+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Fixed new command not to use all params to create a path #1006 
+  (@mlitwiniuk)
 
 ## [2.0.0.beta6] - 2025-04-16